Switched to using EventDispatcher for Comm events instead of the single function the wave-as-client guys did.
Added in a ModeChange set of events.  These only actually work in Stub because wave-as-client doesn't currently implement it.

Made an EditableText component that is either editable or not depending on current Mode (view or edit mode).

Made a quick widget for Hp display info for use in my Hp Gadget.  Doesn't actually do much yet.

Made a widget for switching between Edit and View modes, and added it to both gadgets.